requirements.txt
# Core dependencies
streamlit>=1.28.0
langchain>=0.1.0
langchain-google-genai>=0.1.0
chromadb>=0.4.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# Document processing
PyPDF2>=3.0.0

# Optional: For better PDF support
# pypdf>=3.0.0  # Alternative to PyPDF2 